Refactor sources, auto-gen RCC file at runtime
