+++ .gitignore
@@ -26,3 +26,5 @@
 *.exe
 *.out
 *.app
+*.qmlc
+*.jsc
