--- app/components
@@ -1,1 +1,0 @@
-../monero-gui/components
