--- app/images
@@ -1,1 +1,0 @@
-../monero-gui/images
