--- app/js
@@ -1,1 +1,0 @@
-../monero-gui/js
