--- app/pages
@@ -1,1 +1,0 @@
-../monero-gui/pages
