--- app/wizard
@@ -1,1 +1,0 @@
-../monero-gui/wizard
