+++ qml-xmr/src/mocks/Wallet.h
@@ -1,0 +1,14 @@
+#ifndef WALLET_H
+#define WALLET_H
+
+#include <QObject>
+#include <QTime>
+#include <QMutex>
+#include <QList>
+
+
+class Wallet : public QObject {
+    Q_OBJECT
+};
+
+#endif // WALLET_H
