+++ qml-xmr/src/oscursor.cpp
