+++ qml-xmr/src/window.h
@@ -1,0 +1,39 @@
+#ifndef QML_XMR_WINDOW_H
+#define QML_XMR_WINDOW_H
+
+#include <time.h>
+#include <stdlib.h>
+
+#include <QtCore/qabstractanimation.h>
+#include <QtCore/qdir.h>
+#include <QtCore/qmath.h>
+#include <QtCore/qdatetime.h>
+#include <QtCore/qpointer.h>
+#include <QtCore/qscopedpointer.h>
+#include <QtCore/qtextstream.h>
+
+#include <QtGui/QGuiApplication>
+
+#include <QtQml/qqml.h>
+#include <QtQml/qqmlengine.h>
+#include <QtQml/qqmlcomponent.h>
+#include <QtQml/qqmlcontext.h>
+
+#include <QtQuick/qquickitem.h>
+#include <QtQuick/qquickview.h>
+
+#include <QtWidgets/QApplication>
+#include <QtWidgets/QFileDialog>
+
+#include <QtCore/QTranslator>
+#include <QtCore/QLibraryInfo>
+
+#include <QThread>
+#include <QFileSystemWatcher>
+
+QQuickWindow * createWindow(QQmlEngine *engine_handler,
+                            QQuickView *qxView, QPointer<QQmlComponent> component,
+                            StartOptions options,
+                            QQuickWindow *window = NULL);
+
+#endif //QML_XMR_WINDOW_H
